+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### Unreleased
+* Add `metadata` field to `JobStatus`
 * Add `interval_minutes` field in Scheduler booking config
 * Fixed `Event.originalStartTime` type to be `Date`
 * Fixed `SchedulerBookingOpeningHours.accountId` deserialization